Made exclusion methods private
