+++ src/main/lib/file-utils.js
@@ -1,7 +1,6 @@
 const fs = require('fs');
 const path = require('path');
 const glob = require('glob');
-const exclusionUtils = require('./exclusion-utils');
 const matcher = require('matcher');
 
 const readJson = (filename, processFn) => {
@@ -34,7 +33,7 @@ const exists = (path) => fs.existsSync(path);
 const getJsonFilePaths = (directory, exclusions = []) => {
   const paths = glob.sync(directory + '/**/*.json');
   const relativePaths = toRelativePaths(paths, directory);
-  exclusions = exclusions.map(exclusion => exclusionUtils.prepareExclusion(exclusion));
+  exclusions = exclusions.map(exclusion => prepareExclusion(exclusion));
   return exclusions.length === 0 ? relativePaths : relativePaths.filter(
     path => !exclusions.some(exclusion => matcher.isMatch(path, exclusion)));
 };
@@ -42,6 +41,31 @@ const getJsonFilePaths = (directory, exclusions = []) => {
 function toRelativePaths (array, root) {
   return array.map(
     file => path.relative(root, file));
+}
+
+function isDirExclusion (exclusion) {
+  return exclusion.indexOf('.') === -1;
+}
+
+function prepareExclusion (exclusion) {
+  let prepared = exclusion;
+  if (isDirExclusion(exclusion)) {
+    if (prepared.charAt(prepared.length - 1) !== '/') {
+      prepared = prepared + '/';
+    }
+    if (exclusion.indexOf('*') !== 0) {
+      prepared = '*' + prepared;
+    }
+    if (exclusion.lastIndexOf('*') !== exclusion.length - 1) {
+      prepared = prepared + '*';
+    }
+  } else {
+    if (exclusion.indexOf('*') !== 0) {
+      prepared = '*' + prepared;
+    }
+  }
+
+  return prepared;
 }
 
 module.exports = {
